% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{check_input}
\alias{check_input}
\title{Title}
\usage{
check_input(
  acc_df,
  acc_folder,
  descriptives_df,
  timeCols = "timestamp",
  accCol = "acc",
  factor_split,
  factor_along
)
}
\arguments{
\item{timeCols}{}

\item{accCol}{}

\item{df}{}

\item{datadir}{}

\item{epoch}{}

\item{acc_metric}{}
}
\description{
Title
}
