% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cutpoints.R
\name{get_cutpoints}
\alias{get_cutpoints}
\title{Title}
\usage{
get_cutpoints(
  acc_df = NULL,
  acc_folder = NULL,
  descriptives_df = NULL,
  factor_split = "RIAGENDR",
  factor_along = "RIDAGEYR",
  timeCols = "timestamp",
  accCol = "acc",
  acc_metric = ""
)
}
\arguments{
\item{timeCols}{}

\item{accCol}{}

\item{acc_metric}{}

\item{df}{}

\item{datadir}{}
}
\description{
Title
}
