% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_target_intensity.R
\name{get_target_intensity}
\alias{get_target_intensity}
\title{Title}
\usage{
get_target_intensity(
  data = NULL,
  acc_metric = "MaxMax5min",
  factor_split = c("sex"),
  factor_fit = "age",
  percentile = 80
)
}
\arguments{
\item{data}{}

\item{acc_metric}{}

\item{factor_split}{}

\item{factor_fit}{}

\item{percentile}{}
}
\description{
Title
}
