% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isNHANES.R
\name{isNHANES}
\alias{isNHANES}
\title{Checks whether provided data frame is NHANES data}
\usage{
isNHANES(df)
}
\arguments{
\item{df}{Data frame with, at least, ID, timestamp, and acceleration metric.}
}
\value{
Logical indicating whether the data frame provided corresponds with
NHANES data.
}
\description{
Based on the column names, if SEQN, PAXHOUR, PACMINUT, and PAXINTEN are
column names in the data frame provided. Then, the data is assumed to be
NHANES data.
}
