% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxth.R
\name{maxth}
\alias{maxth}
\title{Maximum X Minutes}
\usage{
maxth(x, k, epoch, consecutive = FALSE)
}
\arguments{
\item{x}{Numeric vector with acceleration values per epoch.}

\item{k}{Number indicating the kth maximum minute in the day.}

\item{epoch}{Number indicating the epoch length in seconds.}

\item{consecutive}{Logical indicating whether k minutes should be consecutive or not.}
}
\value{
Max th minute in the day
}
\description{
Maximum X Minutes
}
