% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popmetrics-package.R
\docType{package}
\name{popmetrics-package}
\alias{popmetrics}
\alias{popmetrics-package}
\title{popmetrics: Age- and Sex-Specific Accelerometer Cut-Points}
\description{
Tools to derive age- and sex-specific cut-points based on accelerometer population-standardized values.
}
\author{
\strong{Maintainer}: Jairo H. Migueles \email{jairo@jhmigueles.com} (\href{https://orcid.org/0000-0003-0366-6935}{ORCID})

}
\keyword{internal}
